Sequential Stages
Stages in Declarative Pipeline may declare a list of nested stages to be run within them in sequential order. Note that a stage must have one and only one of steps, parallel, or stages, the last for sequential stages. It is not possible to nest a parallel block within a stage directive if that stage directive is nested within a parallel block itself. However, a stage directive within a parallel block can use all other functionality of a stage, including agent, tools, when, etc.